LayoutTests/fast/dom/HTMLDialogElement/dialog-show-modal-expected.txt

Tests that showModal() performs the steps specified in the HTML spec. bug 97425

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedStyle.getPropertyValue('display') is "none"
PASS computedStyle.getPropertyValue('display') is "block"
"If dialog already has an open attribute, then throw an InvalidStateError exception."
PASS dialog.showModal(); threw exception Error: InvalidStateError: DOM Exception 11.
PASS computedStyle.getPropertyValue('display') is "none"
"If dialog is not in a Document, then throw an InvalidStateError exception."
PASS dialog.showModal(); threw exception Error: InvalidStateError: DOM Exception 11.
PASS dialog.open is false
Although the document is not attached to any pages, showModal() should execute as normal.
PASS dialog.open is true
PASS successfullyParsed is true

TEST COMPLETE
